Chapter start, end bounds mandatory
Each book must have at least one chapter. If a book is not
chapterised, a dummy chapter spanning the entire book length is
created copying the book metadata. Each chapter could be a segment
of a single m4b book file or could be an independent file.
That way, a 'player' is always a 'chapter player'.