Attach PlaybackConnection to activity instead of application
PlaybackConnection was always meant to have its lifecycle tied to stop/start of MainActivity.
This change became necessary for notifications to work properly, since when a notification
is tapped, a new MainActivity instance is launched and onStop for the old is called AFTER
onStart for the new. The results in singleton PlaybackConnection being released just after
notification => activity launch. Have to look into using @ActivityRetainedScoped