Link activity lifecycle to PlaybackService connection
Notifications cause an interesting case where a new activity is created on tapping of notifications,
and the onDestroy of the old activity is fired with some delay. Since a Singleton is used to hold
the service connection, the delayed onDestroy clears a newly created connection.
Singletons are always bad.
However, now we create a new connection on every UI entry